Ensure returned values are always float
The secs variable can either be a float or an integer, and depending on
this the final result of this function can be of either type too. The
documentation says this function always returns a float though, and it's
good to stay consistent, so this commit ensures the result is always of
type float.

This problem was originally reported in #87.

+++ src/ngamsCore/ngamsLib/ngamsCore.py
@@ -640,11 +640,11 @@ def isoTime2Secs(isoTime):
     timeVector = timePart.split(":")
     hours = int(timeVector[0])
     mins = int(timeVector[1])
-    secs = 0
+    secs = 0.0
     if len(timeVector) > 2:
         secs = float(timeVector[2])
 
-    return days*24*3600 + hours*3600 + mins*60 + secs
+    return float(days*24*3600 + hours*3600 + mins*60) + secs
 
 
 def getBoolean(val):
